templates: Upgrade gradle plugins
For new workload created with `wa create workload -k apkuiauto` gradle
was failing with:

    No toolchains found in the NDK toolchains folder for ABI with
    prefix mips64el-linux-android

NDK stopped shipping the mips toolchain years ago.  Upgrade gradle so
that it doesn't depend on deprecated stuff in the NDK.

The new gradle does not support absolute paths for the
outputfile (which is now called outputFileName, instead of
outputFile) so fix that as well.